refactor: add new modules, extract tests Motivation: The file sizes are growing very large with thousands of LOC.
